Updated package.json and moved Button component to components file
