+++ src/components/TrackingSection/TrackingSection.scss
@@ -22,7 +22,7 @@ body {
 
 .tracking_section {
     padding: 10rem 0;
-    background: linear-gradient(to bottom, rgba(22, 22, 22, 0.6) 0%, rgba(22, 22, 22, 0.8) 75%, #161616 100%), url("../../../assets/images/tracking-background.jpg");
+    background: linear-gradient(to bottom, rgba(22, 22, 22, 0.6) 0%, rgba(22, 22, 22, 0.8) 75%, #161616 100%), url("../../assets/images/tracking-background.jpg");
     background-position: center;
     background-repeat: no-repeat;
     background-attachment: scroll;
