+++ src/screen/Login.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Login from '../components/RegisterForm/Login/Login'
+import Login from '../components/Login/Login'
 import Footer from '../components/Footer/Footer'
 const Login1 = () => {
