+++ src/screen/Tracking.js
@@ -1,5 +1,5 @@
 import React from "react";
-import TrackingSection from "../components/TrackingSection/TrackingSection/TrackingSection";
+import TrackingSection from "../components/TrackingSection/TrackingSection";
 import Footer from '../components/Footer/Footer'
 
 const Tracking = () => {
